[ch37311] Move common components like language-dropdown, country-dropdown, organization-dropdown, support button, page footer to etools-modules-common, etools-unicef
