+++ package.json
@@ -37,10 +37,10 @@
   "dependencies": {
     "@types/lodash-es": "^4.17.7",
     "@shoelace-style/shoelace": "2.10.0",
-    "@unicef-polymer/etools-modules-common": "^5.0.0-rc.52",
+    "@unicef-polymer/etools-modules-common": "^5.0.0-rc.53",
     "@unicef-polymer/etools-piwik-analytics": "4.0.1-rc.3",
     "@unicef-polymer/etools-types": "^1.11.17",
-    "@unicef-polymer/etools-unicef": "^1.0.0-rc.92",
+    "@unicef-polymer/etools-unicef": "^1.0.0-rc.96",
     "@unicef-polymer/etools-utils": "^2.0.0-rc.11",
     "dayjs": "^1.11.10",
     "dexie": "^3.2.4",
